+++ toys-core/lib/toys/input_file.rb
@@ -40,18 +40,19 @@ module Toys::InputFile # rubocop:disable Style/ClassAndModuleChildren
   end
 
   ## @private
-  def self.evaluate(tool_class, remaining_words, path, data_finder)
+  def self.evaluate(tool_class, remaining_words, source)
     namespace = ::Module.new
     namespace.module_eval do
       include ::Toys::Tool::Keys
       @tool_class = tool_class
     end
+    path = source.source_path
     basename = ::File.basename(path).tr(".-", "_").gsub(/\W/, "")
     name = "M#{namespace.object_id}_#{basename}"
     str = build_eval_string(name, ::IO.read(path))
     if str
       const_set(name, namespace)
-      ::Toys::DSL::Tool.prepare(tool_class, remaining_words, path, data_finder) do
+      ::Toys::DSL::Tool.prepare(tool_class, remaining_words, source) do
         ::Toys::ContextualError.capture_path("Error while loading Toys config!", path) do
           # rubocop:disable Security/Eval
           eval(str, __binding, path, 0)