+++ toys-core/test/test_loader.rb
@@ -51,7 +51,7 @@ describe Toys::Loader do
 
   describe "configuration block" do
     it "loads tools" do
-      loader.add_block(path: "test block") do
+      loader.add_block(name: "test block") do
         tool "tool-1" do
           desc "block tool-1 description"
         end
@@ -59,17 +59,17 @@ describe Toys::Loader do
       tool, remaining = loader.lookup(["tool-1"])
       assert_equal("block tool-1 description", tool.desc.to_s)
       assert_equal(true, tool.definition_finished?)
-      assert_equal("test block", tool.source_path)
+      assert_equal("test block", tool.source_name)
       assert_equal([], remaining)
     end
 
     it "loads multiple blocks" do
-      loader.add_block(path: "test block 1") do
+      loader.add_block(name: "test block 1") do
         tool "tool-1" do
           desc "block 1 tool-1 description"
         end
       end
-      loader.add_block(path: "test block 2") do
+      loader.add_block(name: "test block 2") do
         tool "tool-1" do
           desc "block 2 tool-1 description"
         end
@@ -80,12 +80,12 @@ describe Toys::Loader do
       tool, remaining = loader.lookup(["tool-1"])
       assert_equal("block 1 tool-1 description", tool.desc.to_s)
       assert_equal(true, tool.definition_finished?)
-      assert_equal("test block 1", tool.source_path)
+      assert_equal("test block 1", tool.source_name)
       assert_equal([], remaining)
       tool, remaining = loader.lookup(["tool-2"])
       assert_equal("block 2 tool-2 description", tool.desc.to_s)
       assert_equal(true, tool.definition_finished?)
-      assert_equal("test block 2", tool.source_path)
+      assert_equal("test block 2", tool.source_name)
       assert_equal([], remaining)
     end
   end
@@ -147,6 +147,7 @@ describe Toys::Loader do
       tool, remaining = loader.lookup(["tool-1"])
       assert_equal("normal tool-1 short description", tool.desc.to_s)
       assert_equal([], remaining)
+      assert_equal(cases_dir, tool.context_directory)
     end
 
     it "finds a subtool directly defined" do
@@ -154,6 +155,7 @@ describe Toys::Loader do
       assert_equal("normal tool-1-3 short description", tool.desc.to_s)
       assert_equal(["namespace-1", "tool-1-3"], tool.full_name)
       assert_equal([], remaining)
+      assert_equal(cases_dir, tool.context_directory)
     end
 
     it "finds a namespace directly defined" do
@@ -283,11 +285,13 @@ describe Toys::Loader do
     it "gets an item from a root-level file include" do
       tool, _remaining = loader.lookup(["namespace-1", "tool-1-1"])
       assert_equal("file tool-1-1 short description", tool.desc.to_s)
+      assert_equal(cases_dir, tool.context_directory)
     end
 
     it "gets an item from non-root-level include" do
       tool, _remaining = loader.lookup(["namespace-0", "namespace-1", "tool-1-1"])
       assert_equal("normal tool-1-1 short description", tool.desc.to_s)
+      assert_equal(cases_dir, tool.context_directory)
     end
 
     it "does not load an include if not needed" do
@@ -370,7 +374,7 @@ describe Toys::Loader do
     end
   end
 
-  describe "data finder" do
+  describe "with data directory" do
     let(:finding_loader) {
       Toys::Loader.new(index_file_name: ".toys.rb",
                        data_directory_name: ".data")