+++ toys/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### 0.5.1 / TBD
 
+* IMPROVED: Context directory is now available to tools.
+
 ### 0.5.0 / 2018-10-07
 
 * ADDED: Period and colon are recognized as tool path delimiters.